Tire distance variables converted to double

+++ Speed/p_mobj_h.pas
@@ -494,9 +494,9 @@ type
     currPath: integer;
     carinfo: integer;
     nextparticle: integer;  // Next tic that allows particle
-    prevtiredistance: array[0..3] of integer;
-    nexttiredistance: array[0..3] of integer;
-    tiredistance: array[0..3] of integer; // Distance each tire has travel
+    prevtiredistance: array[0..3] of double;
+    nexttiredistance: array[0..3] of double;
+    tiredistance: array[0..3] of double; // Distance each tire has travel
     prevtireangle: integer;
     nexttireangle: integer;
     tireangle: integer;
